WIP: Throw a special error when gql handler not found
